Criando uma classe de erros

+++ ignite/Projeto2-igniteTeams/src/storage/group/groupCreate.ts
@@ -1,15 +1,21 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { groupsGetAll } from "@storage/groupsGetAll";
+import { AppError } from "@utils/AppError";
 
+import { groupsGetAll } from "@storage/groupsGetAll";
 import { GROUP_COLECTION } from "@storage/storageConfig";
 
 export async function groupCreate(newGroup: string) {
     try {
         const storedGroups = await groupsGetAll();
 
-        const storage = JSON.stringify([...storedGroups, newGroup])
+        const groupAlreadyExists = storedGroups.includes(newGroup);
 
-        await AsyncStorage.setItem(GROUP_COLECTION, storage)
+        if (groupAlreadyExists) {
+            throw new AppError('Já existe um grupo cadastrado com esse nome.')
+        }
+
+        const storage = JSON.stringify([...storedGroups, newGroup]);
+        await AsyncStorage.setItem(GROUP_COLECTION, storage);
     } catch (error) {
         throw error;
     }
